+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## [Unreleased]
 ### Added
-- `--force-enac` flag to force ENAC to 1 after calibration
 - Add support for python to 3.9.
 - Update mu_3sl library to version "3.4.3.1.post2".
+- Add encoder configuration file "config\encoders.json".
 
 ## [0.1.2] - 2026-06-02
 ### Added